Simplified BoardingPass construction

+++ src/main/java/com/adventofcode/day5/Controller.java
@@ -15,9 +15,9 @@ public class Controller {
         List<String> input = inputReader.read("puzzle-input/day-5.txt");
         input.stream().map(BoardingPass::new).map(BoardingPass::getSeadId).max(Long::compare).ifPresent(rec -> log.info("Max seat number: {}", rec));
 
-        List<Long> seatIds = input.stream().map(BoardingPass::new).map(BoardingPass::getSeadId).sorted().collect(Collectors.toList());
-        long previousSeatId = 69L;
-        for (Long seatId : seatIds) {
+        List<Integer> seatIds = input.stream().map(BoardingPass::new).map(BoardingPass::getSeadId).sorted().collect(Collectors.toList());
+        int previousSeatId = 69;
+        for (Integer seatId : seatIds) {
             if (previousSeatId != (seatId - 1)) {
                 log.info("My seat number: {}", seatId - 1);
             }
